mapping the errors
