+++ lib/features/home/common/bloc/home_bloc.dart
@@ -3,10 +3,12 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:sentry_flutter/sentry_flutter.dart';
 
 import '../../../../../common/repository/database_repository.dart';
 import '../../../../../core/di/injectable.dart';
 import '../../../../common/data/models/models.dart';
+import '../../../../common/utils/app_util.dart';
 
 part 'home_event.dart';
 part 'home_state.dart';
@@ -26,9 +28,15 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
     Emitter<HomeState> emit,
   ) async {
     emit(HomeFetchingState());
-    var books = await _dbRepo.fetchBooks();
-    var songs = await _dbRepo.fetchSongs();
-    emit(HomeFetchedState(books, songs));
+    try {
+      var books = await _dbRepo.fetchBooks();
+      var songs = await _dbRepo.fetchSongs();
+      emit(HomeFetchedState(books, songs));
+    } catch (e) {
+      logger('Unable to: $e');
+      Sentry.captureException(e);
+      emit(HomeFailureState('Unable to fetch songs'));
+    }
   }
 
   Future<void> _onFilterData(
@@ -36,8 +44,14 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
     Emitter<HomeState> emit,
   ) async {
     emit(HomeFilteringState());
-    var songs = await _dbRepo.fetchSongs(bid: event.book.bookId!);
-    var likes = await _dbRepo.fetchLikes();
-    emit(HomeFilteredState(event.book, songs, likes));
+    try {
+      var songs = await _dbRepo.fetchSongs(bid: event.book.bookId!);
+      var likes = await _dbRepo.fetchLikes();
+      emit(HomeFilteredState(event.book, songs, likes));
+    } catch (e) {
+      logger('Unable to: $e');
+      Sentry.captureException(e);
+      emit(HomeFailureState('Unable to fetch songs'));
+    }
   }
 }
